+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/simplifications/RecordCompiler.java
@@ -1,4 +1,4 @@
-package com.aws.jverify.verifier.compiler.simplifications.temporary;
+package com.aws.jverify.verifier.compiler.simplifications;
 
 import com.aws.jverify.Modifiable;
 import com.aws.jverify.generated.*;
